# backend/swagger.yaml
openapi: "3.0.0"
info:
  title: McMaster CanIDS
  version: 2.0.0
paths:
...
  /api/user/update:
    post:
      summary: Update existing user
      description: |
        Update an existing user account.
        
        If the `uuid` query parameter is present, this will modify the other user. If the query parameter is not present, this will modify the requesting user.
        
        Restrictions:
          - `standard` can only modify own fields:
            - `name`
            - `uuid`
          - `admin` can modify `admin` and `standard` fields:
            - `name`
            - `uuid`
            - `class`
            - `activated`
          - `admin` can not promote anyone to `superuser`.
          - `superuser` can everyone's fields:
            - `name`
            - `uuid`
            - `class`
            - `activated`
            - `group`
          
        To prevent lockout, an `admin` and `superuser` can not modify their own `class` and `activated`.
        
        While the database is modified in real time, a logged in user may not notice changes for up to 1 minute. The APIs will update in real time. This is due to token caching.
        
      tags:
      - User
      parameters: 
        - in: query
          name: uuid
          schema:
            type: string
          description: |
            UUID of other user to modify.
          required: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: |
                    Name of user
                uuid:
                  type: string
                  description: |
                    Email address of user
                class:
                  type: string
                  description: |
                    Type of user account: `standard`, `admin`, `superuser`
                activated:
                    type: boolean
                    description: |
                      Indicator if account is activated
                group:
                  type: string
                  description: |
                    Group UUID
              required:
                - name
                - uuid
                - class
                - activated
                - group
            example: {
              "name": "Tanner Ryan",
              "uuid": "ryant3@mcmaster.ca",
              "class": "admin",
              "activated": true,
              "group": "175dd231-f547-4342-a33e-c45d5e99aa7e"
            }
      responses:
        '200':
          description: |
            Account successfully updated. Changes will be applied within 1 minute.
          content:
            application/json:
              example: {
                "success": true,
                "message": "The user account has been successfully updated. Changes will be applied within 1 minute."
              }
        '400':
          description: |
            Request parameters are not valid.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Bad request format."
              }
        '401':
          description: |
            User is not authenticated
        '403':
          description: |
            User does not have permissions to perform requested actions.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Standard users can only modify their own account."
              }
        '500':
          description: |
            Internal server error
...
  /api/group/update:
    post:
      summary: Update existing group
      description: |
        Update existing group.
        
        Restrictions:
          - `superuser` is the only class of accounts allowed to update groups.

      tags:
      - Group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                uuid:
                  type: string
                  description: |
                    UUID of group to update
                name:
                  type: string
                  description: |
                    Updated or existing group name
                authorized:
                  type: array
                  description: |
                    Updated or existing group assets
                  items:
                    type: string
              required:
                - uuid
                - name
                - authorized
            example: {
              "uuid": "175dd231-f547-4342-a33e-c45d5e99aa7e",
              "name": "McMaster University",
              "authorized": [
                "asset1",
                "asset2"
              ]
            }
      responses:
        '200':
          description: |
            Group successfully deleted.
          content:
            application/json:
              example: {
                "success": true,
                "message": "Successfully deleted group."
              }
        '400':
          description: |
            Request parameters are not valid.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Bad request format."
              }
        '401':
          description: |
            User is not authenticated
        '403':
          description: |
            User does not have permissions to perform requested actions.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Only a superuser can delete a group."
              }
        '500':
          description: |
            Internal server error
...
  /api/view/update:
    post:
      summary: Upate existing view
      description: |
        Update an existing view.
        
        Restrictions:
          - `admin` can update views belonging to their group.
          - `superuser` can update views belonging to any group.
          
      tags:
      - View
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                uuid:
                  type: string
                  description: |
                    UUID of view to update
                group:
                  type: string
                  description: |
                    Group that view belongs to
                authorized:
                  type: string
                  description: |
                    Authorized asset to populate view with (source index)
                name:
                  type: string
                  description: |
                    Name of view
                class:
                  type: string
                  description: |
                    Type of view: `line`, `bar`, `pie`, `table`
                field:
                  type: string
                  description: |
                    Data field to use for view
                fieldName:
                  type: string
                  description: |
                    Common field name
              required:
                - uuid
                - group
                - authorized
                - name 
                - class
                - field
                - fieldName
            example: {
              "uuid": "70884425-0cab-4840-a99e-ee397a9dbc71",
              "group": "6a7d3a03-5c06-4620-8325-583c8136b17f",
              "authorized": "mcmaster_001",
              "name": "JHE Incoming Packets",
              "class": "line",
              "field": "packets_incoming",
              "fieldName": "Packets"
            }
      responses:
        '200':
          description: |
            View successfully updated.
          content:
            application/json:
              example: {
                "success": true,
                "message": "View successfully updated."
              }
        '400':
          description: |
            Request parameters are not valid.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Bad request format."
              }
        '401':
          description: |
            User is not authenticated
        '403':
          description: |
            User does not have permissions to perform requested actions.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Admin users can not add foreign views."
              }
        '500':
          description: |
            Internal server error
...
  /api/dashboard/update:
    post:
      summary: Update dashboard
      description: |
        Update an existing view.
        
        Restrictions:
          - `admin` can update their group's dashboard
          - `superuser` can update any group's dashboard
      tags:
      - Dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                uuid:
                  type: string
                  description: |
                    Dashboard UUID
                group:
                  type: string
                  description: |
                    UUID of group dashboard belongs to
                name:
                  type: string
                  description: |
                    Dashboard name
                views:
                  type: array
                  description: |
                    List of views in dashboard
                  items:
                    type: string
                    description: |
                      View UUID in dashboard
                sizes:
                  type: array
                  description: |
                    List of sizes corresponding to views: `half` or `full` width
                  items:
                    type: string
                    description: |
                      Size corresponding to UUID in dashboard
              required:
              - uuid
              - group
              - name
              - views
              - sizes
            example: {
              "uuid": "7290169e-9bdf-437b-8dd6-456b17341f05",
              "group": "6a7d3a03-5c06-4620-8325-583c8136b17f",
              "name": "McMaster University Dashboard",
              "views": [
                "70884425-0cab-4840-a99e-ee397a9dbc71",
                "b93b03f2-64ce-45cd-8737-9f99ed885253"
              ],
              "sizes": [
                "full",
                "full"
              ]
            }
      responses:
        '200':
          description: |
            Dashboard successfully updated.
          content:
            application/json:
              example: {
                "success": true,
                "message": "Dashboard successfully updated."
              }
        '400':
          description: |
            Request parameters are not valid.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Bad request format."
              }
        '401':
          description: |
            User is not authenticated
        '403':
          description: |
            User does not have permissions to perform requested actions.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Admin users can not update foreign dashboard."
              }
        '500':
          description: |
            Internal server error